+++ src/ThemeContext.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
 
+import { createThemedDefaultCacheManager } from './creators';
 import {
   ExtractThemeNames,
   ThemeContextProps,
@@ -10,26 +11,42 @@ import {
   Styles,
 } from './types';
 
-export const ThemeContext = React.createContext<ThemeContextValue<any> | null>(
-  null,
-);
+export const ThemeContext = React.createContext<ThemeContextValue<
+  any,
+  any
+> | null>(null);
 
 export const ThemeBaseStylesContext =
   React.createContext<ThemeBaseStylesContextValue<any> | null>(null);
 
 export const ThemeDispatchContext =
-  React.createContext<ThemeDispatchContextValue<any> | null>(null);
+  React.createContext<ThemeDispatchContextValue<any, any> | null>(null);
 
-export function ThemeProvider<T extends Themes, BS extends Styles<BS>>({
+export function ThemeProvider<T extends Themes, BS extends Styles<BS>, P>({
   children,
   initialTheme,
   onThemeChange,
+  onThemeParamsChange,
   themes,
   baseStylesCreator,
-}: ThemeContextProps<T, BS>) {
+  initialThemeParams,
+  styleCacheManager = createThemedDefaultCacheManager<T>(),
+}: ThemeContextProps<T, BS, P>) {
   const [themeName, setThemeName] =
     React.useState<ExtractThemeNames<T>>(initialTheme);
-  const t = React.useMemo(() => themes[themeName], [themes, themeName]);
+  const [themeParams, setThemeParams] = React.useState(initialThemeParams);
+
+  React.useEffect(() => {
+    styleCacheManager.onProviderMount && styleCacheManager.onProviderMount();
+  }, [styleCacheManager]);
+
+  const t = React.useMemo(
+    () =>
+      typeof themes === 'function'
+        ? themes(themeParams as P)[themeName]
+        : themes[themeName],
+    [themes, themeName, themeParams],
+  );
 
   const baseStyles = React.useMemo(
     () => (baseStylesCreator ? baseStylesCreator(t) : null),
@@ -40,14 +57,27 @@ export function ThemeProvider<T extends Themes, BS extends Styles<BS>>({
     (nextTheme: any) => {
       onThemeChange && onThemeChange(nextTheme);
       setThemeName(nextTheme);
+      styleCacheManager.onThemeChange(nextTheme);
     },
-    [onThemeChange],
+    [onThemeChange, styleCacheManager],
+  );
+
+  const changeThemeParams = React.useCallback(
+    (nextParams: P) => {
+      onThemeParamsChange && onThemeParamsChange(nextParams);
+      setThemeParams(nextParams);
+    },
+    [onThemeParamsChange],
   );
 
   return (
-    <ThemeContext.Provider value={{ selectedTheme: themeName, themes, t }}>
+    <ThemeContext.Provider
+      value={{ selectedTheme: themeName, themes, t, themeParams }}
+    >
       <ThemeBaseStylesContext.Provider value={{ baseStyles }}>
-        <ThemeDispatchContext.Provider value={{ setTheme: changeTheme }}>
+        <ThemeDispatchContext.Provider
+          value={{ setTheme: changeTheme, setParams: changeThemeParams }}
+        >
           {children}
         </ThemeDispatchContext.Provider>
       </ThemeBaseStylesContext.Provider>