+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.1.2
 
 - Add support for passing default styles, see [Passing and accessing default styles](./README.md#passing-and-accessing-default-styles)
+- Add support for passing theme params, see [Initialize themes and the theme provider](./README.md#initialize-themes-and-the-theme-provider) and [Change or access theme params](./README.md#change-or-access-theme-params)
 
 ## 0.1.1 and lower